requirements.txt
watchdog
fastapi[standard]
pandas
djongo
pymongo
pymongo[srv]
motor
python-dotenv